requirements.txt
numpy
pandas
scikit-learn
torch
pyyaml
tqdm
rdkit-pypi
matplotlib
shap
seaborn
openpyxl
